Arel::Nodes::Function
