+++ lib/arel/extensions/nodes.rb
@@ -4,22 +4,6 @@
 
 module Arel
   module Nodes
-    Arel::Nodes::Function.class_eval do
-      # postgres only: https://www.postgresql.org/docs/9.5/functions-aggregate.html
-      attr_accessor :orders
-      attr_accessor :filter
-      attr_accessor :within_group
-      attr_accessor :variardic
-
-      def initialize(expr, aliaz = nil)
-        super()
-        @expressions = expr
-        @alias       = aliaz && SqlLiteral.new(aliaz)
-        @distinct    = false
-        @orders      = []
-      end
-    end
-
     # https://www.postgresql.org/docs/9.4/functions-math.html
     class Factorial < Arel::Nodes::Unary
       attr_accessor :prefix
